very verbose solution for textfield submission / validation in imgui
ParseIntField and ParseFloatField will either return the value they're supposed to present or a temporary inputbuffer depending on whether or not they're focused in the imgui. I had to put a lot more code into OnGUI than I'd like since I've had trouble finding a solution that allows use of functions for creating fields. This works well in the editor but its not the most reusable. I'd like to find a better solution should one present itself.